+++ admin/tabs/activity/clienterrors/ClientErrorModel.js
@@ -9,7 +9,7 @@ import moment from 'moment';
 import {XH, HoistModel} from '@xh/hoist/core';
 import {action, observable} from '@xh/hoist/mobx';
 import {LocalStore} from '@xh/hoist/data';
-import {GridModel} from '@xh/hoist/desktop/cmp/grid';
+import {GridModel} from '@xh/hoist/cmp/grid';
 import {fmtDate} from '@xh/hoist/format';
 import {boolCheckCol, compactDateCol} from '@xh/hoist/columns';
 import {usernameCol} from '@xh/hoist/admin/columns';
