+++ admin/tabs/activity/tracking/ActivityGridModel.js
@@ -8,7 +8,7 @@ import moment from 'moment';
 import {XH, HoistModel} from '@xh/hoist/core';
 import {action, observable} from '@xh/hoist/mobx';
 import {LocalStore} from '@xh/hoist/data';
-import {GridModel} from '@xh/hoist/desktop/cmp/grid';
+import {GridModel} from '@xh/hoist/cmp/grid';
 import {fmtDate, numberRenderer} from '@xh/hoist/format';
 import {dateTimeCol} from '@xh/hoist/columns';
 import {usernameCol} from '@xh/hoist/admin/columns';
