+++ cmp/grid/impl/ColChooserModel.js
