+++ cmp/grid/index.js
@@ -9,4 +9,3 @@ export * from './Grid';
 export * from './GridModel';
 export * from './GridStateModel';
 export * from './ColChooserButton';
-export * from './ExportManager';
