+++ desktop/cmp/record/impl/RecordActionButton.js
@@ -3,7 +3,7 @@ import {PropTypes as PT} from 'prop-types';
 import {elemFactory, HoistComponent} from '@xh/hoist/core';
 import {button} from '@xh/hoist/desktop/cmp/button';
 import {RecordAction, Record, StoreSelectionModel} from '@xh/hoist/data';
-import {GridModel} from '@xh/hoist/desktop/cmp/grid';
+import {GridModel} from '@xh/hoist/cmp/grid';
 import {Column} from '@xh/hoist/columns';
 import {first, omit} from 'lodash';
