+++ desktop/cmp/rest/RestGridModel.js
@@ -6,7 +6,7 @@
  */
 import {XH, HoistModel} from '@xh/hoist/core';
 import {action} from '@xh/hoist/mobx';
-import {GridModel} from '@xh/hoist/desktop/cmp/grid';
+import {GridModel} from '@xh/hoist/cmp/grid';
 import {StoreContextMenu} from '@xh/hoist/desktop/cmp/contextmenu';
 import {pluralize} from '@xh/hoist/utils/js';
 import {Icon} from '@xh/hoist/icon/Icon';
