--- mobile/cmp/grid/index.js
@@ -1,9 +1,0 @@
-/*
- * This file belongs to Hoist, an application development toolkit
- * developed by Extremely Heavy Industries (www.xh.io | info@xh.io)
- *
- * Copyright © 2018 Extremely Heavy Industries Inc.
- */
-
-export * from '@xh/hoist/desktop/cmp/grid/Grid';
-export * from '@xh/hoist/desktop/cmp/grid/GridModel';
